roles y permisos

+++ app/Http/Controllers/PermissionController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Permission;
+use Illuminate\Support\Facades\Gate;
 
 class PermissionController extends Controller
 {
@@ -14,7 +15,8 @@ class PermissionController extends Controller
      */
     public function index()
     {
-        //
+        //denegar si quiere entrar escribiendo link
+        abort_if(Gate::denies('permission_index'), 403);
         $permissions=Permission::paginate(5);
         
         return view('permissions.index',compact('permissions'));
@@ -28,6 +30,7 @@ class PermissionController extends Controller
     public function create()
     {
         //
+        abort_if(Gate::denies('permission_create'), 403);
         return view('permissions.create');
     }
 
@@ -54,6 +57,7 @@ class PermissionController extends Controller
     public function show(Permission $permission)
     {
         //
+        abort_if(Gate::denies('permission_show'), 403);
         return view('permissions.show', compact('permission'));
     }
 
@@ -66,7 +70,7 @@ class PermissionController extends Controller
     public function edit(Permission $permission)
     {
         //
-
+        abort_if(Gate::denies('permission_edit'), 403);
         return view('permissions.edit', compact('permission'));
     }
 
@@ -94,6 +98,7 @@ class PermissionController extends Controller
     public function destroy(Permission $permission)
     {
         //
+        abort_if(Gate::denies('permission_delete'), 403);
         $permission->delete();
 
         return redirect()->route('permissions.index');
