+++ app/Http/Controllers/RoleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Permission;
 use Spatie\Permission\Models\Role;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
@@ -16,6 +17,7 @@ class RoleController extends Controller
     public function index()
     {
         //
+        abort_if(Gate::denies('role_index'), 403);
         $roles= Role::paginate(10);
         return view('roles.index', compact('roles'));
     }
@@ -28,6 +30,7 @@ class RoleController extends Controller
     public function create()
     {
         //
+        abort_if(Gate::denies('role_create'), 403);
         $permissions = Permission::all()->pluck('name', 'id');
         // dd($permissions);
         return view('roles.create', compact('permissions'));
@@ -57,6 +60,7 @@ class RoleController extends Controller
     public function show(Role $role)
     {
         //
+        abort_if(Gate::denies('role_show'), 403);
         $role->load('permissions');
         return view('roles.show', compact('role'));
     }
@@ -70,6 +74,7 @@ class RoleController extends Controller
     public function edit(Role $role)
     {
         //
+        abort_if(Gate::denies('role_edit'), 403);
         $permissions = Permission::all()->pluck('name', 'id');
         $role->load('permissions');    
         return view('roles.edit', compact('role', 'permissions'));
@@ -102,6 +107,7 @@ class RoleController extends Controller
     public function destroy(Role $role)
     {
         //
+        abort_if(Gate::denies('role_delete'), 403);
         $role->delete();
         return redirect()->route('roles.index');
     }
